refactor(minor): made more relevent imports & restructured code
